Cleanup checkstyle warnings.
