+++ ehr-server/src/main/java/org/hl7/davinci/ehrserver/requestgenerator/RequestGeneratorController.java
@@ -1,4 +1,4 @@
-package org.hl7.davinci.ehrServer.requestGenerator;
+package org.hl7.davinci.ehrserver.requestgenerator;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Controller;
 public class RequestGeneratorController {
   static final Logger logger = LoggerFactory.getLogger(RequestGeneratorController.class);
 
-  @RequestMapping(value= {"/", "/reqgen", "/index", "/index.html"}, method=RequestMethod.GET)
+  @RequestMapping(value = {"/", "/reqgen", "/index", "/index.html"}, method = RequestMethod.GET)
   public String index() {
     logger.info("RequestGenerator page requested");
     return "index";
